power-exporter-ssl/requirements.txt
prometheus_client
kafka-python
